Add internal-types
