+++ src/models/data-fabric/entity.internal-types.ts
@@ -1,0 +1,15 @@
+// SQL Type Interface
+interface FieldType {
+  name: string;
+}
+
+// Field Definition Interface
+interface FieldDefinition {
+  name: string;
+  sqlType: FieldType;
+}
+
+// Entity Metadata Response Interface
+export interface EntityMetadataResponse {
+  fields: FieldDefinition[];
+}
